Disable USART interrupts.  Use the new send and expect a response function.
Remove redundant varibales and disable interrupts for the USART on PORTC (0). Use the new usart_send_and_expect function to sync up the baud rates of the XMEGA and MC60. Remove redundant code from interrupts. Move the ENVIRONMENT macro to the board configuration file, so it can be used in other files too.