Circular buffer replaces the old infinate length buffer. Remove command helpers. Remove redundant USART helpers.
The old buffer was really slow and would eat up a lot of memory. The new implementation is a FIFO buffer, with a short length, so it saves up on memory. Removed redundant helpers for sending commands, they would also create a lot of objects and allocate a lof of memory, better use strings. USART responses are now handled in an interrupt. This fixes a bug with timing when polling and mingling of RX and TX data. Add a new enum and a function to check the response. General cleanups.